Consistently use `requestLocation()` to get a single location
`startUpdatingLocations()` is not available on watchOS, because the
watch is designed for short interactions and not ongoing updates. Rather
than subscribing to a stream of updates and waiting for a location that
isn't "stale", this change instead just relies on the fact that when
using `requestLocation()`, `locationManager(_:didUpdateLocations:)` will
only be called once.