Resolve an issue where request location would never complete
This change resolves an issue where the location would be updated and a
weather update fetched, but the UI would be stuck, unable to scroll, and
the refresh never appears to complete.

---

When using the `Merge` flatten strategy, the merged signal doesn't
complete until all of the inner producers have completed. It will error
immediately if any inner producer sends an error.

In a success scenario, the `didFailWithError` producer never sends any
value, and never completes. So when merged with the `didUpdateLocations`
producer, the result is a producer that sends a single value and then
never completes (while waiting for the error producer to terminate).

This change introduces a `unify(producers:)` method, that instead of
waiting for all inner producers to complete, immediately is terminated
when _any_ of the inner producers terminates. So once
`didUpdateLocations` sends a single value, the unified producer will
immediately complete instead of waiting for `didFailWithError` to also
complete.