Improve efficiency of weather updates on the watch
The new `requestLocation` API uses its own heuristics to zero in on a
location that best fulfills the minimum required location accuracy. When
requesting the best accuracy, this increases the time taken to return a
location.

This change uses a 1-kilometer radius to speed up location updates,
especially when a very accurate location is time-consuming or impossible
to fetch.

It also stops updating location if the signal producer is interrupted or
errors before completing.